chore: Move tests into subdirectories based on test execution length
