+++ suite/src/__tests__/fast/anchor.test.ts
@@ -3,8 +3,8 @@ import { TileDocument } from '@ceramicnetwork/stream-tile'
 import { afterAll, beforeAll, describe, expect, test } from '@jest/globals'
 import { DateTime } from 'luxon'
 
-import { newCeramic } from '../utils/ceramicHelpers.js'
-import * as helpers from '../utils/dynamoDbHelpers.js'
+import { newCeramic } from '../../utils/ceramicHelpers.js'
+import * as helpers from '../../utils/dynamoDbHelpers.js'
 
 const ComposeDbUrls = String(process.env.COMPOSEDB_URLS).split(',')
 
@@ -44,8 +44,8 @@ describe('anchor', () => {
             `Test failed. StreamID: ${tile.id.toString()}, state:\n${JSON.stringify(
               StreamUtils.serializeState(tile.state),
               null,
-              2,
-            )}`,
+              2
+            )}`
           )
 
           // If the anchoring failed, we don't want to leave this stream in the database,
