+++ suite/src/__tests__/fast/garbage-collect.test.ts
@@ -4,8 +4,8 @@ import { afterAll, beforeAll, describe, test } from '@jest/globals'
 // Internal dependencies
 import CeramicClient from '@ceramicnetwork/http-client'
 import { StreamID } from '@ceramicnetwork/streamid'
-import { newCeramic } from '../utils/ceramicHelpers.js'
-import * as helpers from '../utils/dynamoDbHelpers.js'
+import { newCeramic } from '../../utils/ceramicHelpers.js'
+import * as helpers from '../../utils/dynamoDbHelpers.js'
 
 // Environment variables
 const ComposeDbUrls = String(process.env.COMPOSEDB_URLS).split(',')
