+++ suite/src/__tests__/fast/helpers.test.ts
@@ -1,13 +1,13 @@
 import { StreamID } from '@ceramicnetwork/streamid'
 import {afterAll, beforeAll, describe, expect, test} from '@jest/globals'
 
-import * as helpers from '../utils/dynamoDbHelpers.js'
+import * as helpers from '../../utils/dynamoDbHelpers.js'
 
 const streamID0 = StreamID.fromString(
-  'kjzl6cwe1jw149mj7gw90a40vf9qoipv1gthagyzxzjrg8pp9htbdbvu2er1wan',
+  'kjzl6cwe1jw149mj7gw90a40vf9qoipv1gthagyzxzjrg8pp9htbdbvu2er1wan'
 )
 const streamID1 = StreamID.fromString(
-  'kjzl6cwe1jw149zfreoiia0ayfbwxnc3qizfyouw1c6mp9e3ah3smmhqfrppm0z',
+  'kjzl6cwe1jw149zfreoiia0ayfbwxnc3qizfyouw1c6mp9e3ah3smmhqfrppm0z'
 )
 
 /**
